Directly call GpuExecutable::ExecuteThunks rather than RunAsync. This allows us to take a vector<ShapeTree<pair<raw_buffer, is_donated_bool>>> and return ShapeTree<raw_buffer> and handle aliasing properly for aliased types.
Before this CL, aliased types would be converted to and from ScopedShapedBuffer which means that foreign memory types would lose their attached cleanup callback. Recently this was changed to simply error if the user tries to donate foreign memory types (created via CreateViewOfDeviceBuffer), but this CL makes it supported.

This is also needed to properly track ownership for raw buffer support.

PiperOrigin-RevId: 731914236